requirements.txt
asgiref==3.8.1
Django==5.2.1
django-bootstrap5==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-markdownx==4.0.9
django-shortuuidfield==0.1.3
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.8
packaging==25.0
pillow==11.2.1
pytz==2025.2
PyYAML==6.0.2
setuptools==80.4.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
wheel==0.45.1
